requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
black==25.1.0
celery==5.4.0
certifi==2025.4.26
cfgv==3.4.0
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.1
django-htmx==1.17.3
djangorestframework==3.15.2
filelock==3.18.0
flake8==7.2.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.10
idna==3.10
kombu==5.5.3
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
pre_commit==4.2.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pycodestyle==2.13.0
pyflakes==3.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-telegram-bot==21.5
PyYAML==6.0.2
redis==5.0.8
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
